// OnParDev.MyMcp.Api/ClientApp/src/api/models/ConfigurationResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApiConfiguration } from './ApiConfiguration';
import type { ClerkConfiguration } from './ClerkConfiguration';
import type { FeatureConfiguration } from './FeatureConfiguration';
export type ConfigurationResponse = {
    clerk?: ClerkConfiguration;
    api?: ApiConfiguration;
    features?: FeatureConfiguration;
};